requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0

# ACP (Agent Communication Protocol) – IBM acp-sdk
acp-sdk>=1.0.0
httpx>=0.27.0

# LangChain stack
langchain>=0.2.0
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.3
langchain-groq>=0.1.0

# Vector store & embeddings
chromadb>=0.5.0
sentence-transformers>=2.7.0

# Document loaders
unstructured[pdf,docx,xlsx]>=0.13.0
langchain-text-splitters>=0.2.0

# Utilities
pydantic>=2.0.0
python-multipart>=0.0.6
openpyxl>=3.1.0
requests>=2.31.0
